# Optimize BlockAnalysisVisitor