--- a/src/Phan/AST/Visitor/Element.php
+++ b/src/Phan/AST/Visitor/Element.php
@@ -131,14 +131,17 @@ class Element
     /**
      * Accepts a visitor that differentiates on the kind value
      * of the AST node.
+     *
+     * NOTE: This was turned into a static method for performance
+     * because it was called extremely frequently.
      */
-    public function acceptKindVisitor(KindVisitor $visitor)
+    public static function acceptNodeAndKindVisitor(Node $node, KindVisitor $visitor)
     {
-        $fn_name = self::VISIT_LOOKUP_TABLE[$this->node->kind] ?? null;
+        $fn_name = self::VISIT_LOOKUP_TABLE[$node->kind] ?? null;
         if (is_string($fn_name)) {
-            return $visitor->{$fn_name}($this->node);
+            return $visitor->{$fn_name}($node);
         } else {
-            Debug::printNode($this->node);
+            Debug::printNode($node);
             assert(false, 'All node kinds must match');
         }
     }