--- a/src/Phan/BlockAnalysisVisitor.php
+++ b/src/Phan/BlockAnalysisVisitor.php
@@ -2,6 +2,7 @@
 namespace Phan;
 
 use Phan\AST\AnalysisVisitor;
+use Phan\AST\Visitor\Element;
 use Phan\Analysis\ConditionVisitor;
 use Phan\Analysis\ContextMergeVisitor;
 use Phan\Analysis\PostOrderAnalysisVisitor;
@@ -128,14 +129,44 @@ class BlockAnalysisVisitor extends AnalysisVisitor {
 
             // Step into each child node and get an
             // updated context for the node
-            $context = (new BlockAnalysisVisitor(
-                $this->code_base, $context, $node, $this->depth + 1
-            ))($child_node);
+            $context = $this->analyzeAndGetUpdatedContext($context, $node, $child_node);
         }
 
         $context = $this->postOrderAnalyze($context, $node);
 
         return $context;
+    }
+
+    /**
+     * This is an abstraction for getting a new, updated context for a child node.
+     *
+     * Effectively the same as (new BlockAnalysisVisitor(..., $context, $node, ..., $depth + 1, ...)($child_node))
+     * but is much less repetitive and verbose, and slightly more efficient.
+     *
+     * @param Context $context - The original context for $node, before analyzing $child_node
+     *
+     * @param Node $node - The parent node of $child_node
+     *
+     * @param Node $child_node - The node which will be analyzed to create the updated context.
+     *
+     * @return Context (The unmodified $context, or a different Context instance with modifications)
+     */
+    private function analyzeAndGetUpdatedContext(Context $context, Node $node, Node $child_node) : Context
+    {
+        // Modify the original object instead of creating a new BlockAnalysisVisitor.
+        // this is slightly more efficient, especially if a large number of unchanged parameters would exist.
+        $old_context = $this->context;
+        $old_parent_node = $this->parent_node;
+        $old_depth = $this->depth++;
+        $this->context = $context;
+        $this->parent_node = $node;
+        try {
+            return Element::acceptNodeAndKindVisitor($child_node, $this);
+        } finally {
+            $this->context = $old_context;
+            $this->parent_node = $old_parent_node;
+            $this->depth = $old_depth;
+        }
     }
 
     /**
@@ -205,9 +236,7 @@ class BlockAnalysisVisitor extends AnalysisVisitor {
 
             // Step into each child node and get an
             // updated context for the node
-            $child_context = (new BlockAnalysisVisitor(
-                $this->code_base, $child_context, $node, $this->depth + 1
-            ))($child_node);
+            $child_context = $this->analyzeAndGetUpdatedContext($child_context, $node, $child_node);
 
             // TODO(Issue #406): We can improve analysis of `if` blocks by using
             // a BlockExitStatusChecker to avoid propogating invalid inferences.
@@ -255,24 +284,22 @@ class BlockAnalysisVisitor extends AnalysisVisitor {
 
         $condition_node = $node->children['cond'];
         if ($condition_node && $condition_node instanceof Node) {
-            $context = (new BlockAnalysisVisitor(
-                $this->code_base,
+            $context = $this->analyzeAndGetUpdatedContext(
                 $context->withLineNumberStart($condition_node->lineno ?? 0),
                 $node,
-                $this->depth + 1
-            ))($condition_node);
+                $condition_node
+            );
         }
 
         if ($stmts_node = $node->children['stmts']) {
             if ($stmts_node instanceof Node) {
-                $context = (new BlockAnalysisVisitor(
-                    $this->code_base,
+                $context = $this->analyzeAndGetUpdatedContext(
                     $context->withScope(
                         new BranchScope($context->getScope())
                     )->withLineNumberStart($stmts_node->lineno ?? 0),
                     $node,
-                    $this->depth + 1
-                ))($stmts_node);
+                    $stmts_node
+                );
             }
         }
 
@@ -348,9 +375,7 @@ class BlockAnalysisVisitor extends AnalysisVisitor {
 
             // Step into each child node and get an
             // updated context for the node
-            $child_context = (new BlockAnalysisVisitor(
-                $this->code_base, $child_context, $node, $this->depth + 1
-            ))($child_node);
+            $child_context = $this->analyzeAndGetUpdatedContext($child_context, $node, $child_node);
 
             $child_context_list[] = $child_context;
         }
@@ -444,9 +469,7 @@ class BlockAnalysisVisitor extends AnalysisVisitor {
             // Step into each child node and get an
             // updated context for the node
             // (e.g. there may be assignments such as '($x = foo()) ? $a : $b)
-            $context = (new BlockAnalysisVisitor(
-                $this->code_base, $context, $node, $this->depth + 1
-            ))($cond_node);
+            $context = $this->analyzeAndGetUpdatedContext($context, $node, $cond_node);
 
             // TODO: false_context once there is a NegatedConditionVisitor
             $true_context = (new ConditionVisitor(
@@ -462,18 +485,14 @@ class BlockAnalysisVisitor extends AnalysisVisitor {
         // $cond_node is the (already processed) value for truthy.
         if ($true_node instanceof Node) {
             if ($this->should_visit_everything || Analysis::shouldVisit($true_node)) {
-                $child_context = (new BlockAnalysisVisitor(
-                    $this->code_base, $true_context, $node, $this->depth + 1
-                ))($true_node);
+                $child_context = $this->analyzeAndGetUpdatedContext($true_context, $node, $true_node);
                 $child_context_list[] = $child_context;
             }
         }
 
         if ($false_node instanceof Node) {
             if ($this->should_visit_everything || Analysis::shouldVisit($false_node)) {
-                $child_context = (new BlockAnalysisVisitor(
-                    $this->code_base, $context, $node, $this->depth + 1
-                ))($false_node);
+                $child_context = $this->analyzeAndGetUpdatedContext($context, $node, $false_node);
                 $child_context_list[] = $child_context;
             }
         }