Move recursion into BlockAnalysisVisitor->recurse()
- It was repetitive and required changing multiple invocations to add extra parameters
- If BlockAnalysisVisitor's child visitors ended up having a lot of properties that didn't vary (e.g. if blockexitstatus was added, more config values were added), copying would be inefficient
- Speed self-analysis from 3.90 to 3.80 seconds

Convert acceptKindVisitor into a static method
- It is called extremely frequently.
  (Leave the other methods alone)

PHP can cache the memory address of a static method,
but I don't think it can cache the memory address of an instance method.
Also, invoking the constructor and creating a temporary object adds
extra work.

Time savings for self-analysis (around 4%)

- 3.9s originally
- 3.8s after implementing recurse()
- 3.7s after converting acceptNodeAndKindVisitor to a static method.